predictor_ML/requirements.txt
beautifulsoup4==4.9.1
dateparser==1.1.6
jdatetime==3.6.2
python-dateutil==2.8.1
python-espncricinfo==0.4.1
pytz==2020.1
regex==2020.6.8
requests>=2.20.0
pyyaml
six==1.15.0
umalqurra==0.2
wheel==0.38.0